requirements.txt
altair==4.2.0
appnope==0.1.3
asttokens==2.0.8
attrs==22.1.0
backcall==0.2.0
beautifulsoup4==4.11.1
blinker==1.5
bokeh==2.4.3
cachetools==5.2.0
certifi==2022.6.15.1
charset-normalizer==2.1.1
click==8.1.3
commonmark==0.9.1
cycler==0.11.0
debugpy==1.6.3
decorator==5.1.1
docopt==0.6.2
emot==3.1
entrypoints==0.4
executing==1.1.0
filelock==3.8.0
fonttools==4.37.1
gitdb==4.0.9
GitPython==3.1.27
idna==3.3
importlib-metadata==4.12.0
ipykernel==6.15.3
ipython==8.5.0
ipywidgets==8.0.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.16.0
jupyter-core==4.11.1
jupyter_client==7.3.5
jupyterlab-widgets==3.0.3
kiwisolver==1.4.4
lxml==4.9.1
MarkupSafe==2.1.1
matplotlib==3.5.3
matplotlib-inline==0.1.6
nest-asyncio==1.5.5
nltk==3.7
numpy==1.23.3
packaging==21.3
pandas==1.4.4
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
pipreqs==0.4.11
plotly==5.10.0
prompt-toolkit==3.0.31
protobuf==3.20.1
psutil==5.9.2
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==9.0.0
pydeck==0.7.1
Pygments==2.13.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.18.1
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
pyzmq==24.0.1
regex==2022.9.13
requests==2.28.1
rich==12.5.1
scikit-learn==1.1.2
scipy==1.9.1
seaborn==0.12.0
semver==2.13.0
six==1.16.0
sklearn==0.0
smmap==5.0.0
snscrape==0.4.3.20220106
soupsieve==2.3.2.post1
stack-data==0.5.1
streamlit==1.12.2
streamlit-lottie==0.0.3
tenacity==8.1.0
textblob==0.17.1
threadpoolctl==3.1.0
toml==0.10.2
toolz==0.12.0
tornado==6.2
tqdm==4.64.1
traitlets==5.4.0
typing_extensions==4.3.0
tzdata==2022.2
tzlocal==4.2
urllib3==1.26.12
validators==0.20.0
wcwidth==0.2.5
widgetsnbextension==4.0.3
wordcloud==1.8.2.2
yarg==0.1.9
zipp==3.8.1
